refactor(abw): decompose prompt2blog pipeline
