+++ apps/ai-blog-writer/apps/backend/app/features/prompt2blog/graph/__init__.py
@@ -1,11 +1,5 @@
 """LangGraph execution for Prompt2Blog."""
 
-from .runner import (
-    run_prompt2blog_full_graph,
-    run_prompt2blog_pipeline_v2_graph,
-)
+from .runner import run_prompt2blog_stage_graph
 
-__all__ = [
-    "run_prompt2blog_full_graph",
-    "run_prompt2blog_pipeline_v2_graph",
-]
+__all__ = ["run_prompt2blog_stage_graph"]
