+++ apps/ai-blog-writer/apps/backend/app/features/prompt2blog/stages/__init__.py
@@ -1,0 +1,1 @@
+"""Prompt2Blog pipeline stages."""
