+++ apps/ai-blog-writer/apps/backend/tests/test_langgraph_trace_result_routes.py
@@ -3,6 +3,8 @@ import json
 
 import app.features.prompt2blog.routes as prompt2blog_routes
 import app.features.url2blog.routes as url2blog_routes
+from app.features.prompt2blog import observability as prompt2blog_observability
+from app.features.prompt2blog.api import runs as prompt2blog_runs
 
 
 def _response_json(response) -> dict:
@@ -17,12 +19,12 @@ def test_prompt2blog_result_includes_langgraph_trace_metadata(monkeypatch):
     }
 
     monkeypatch.setattr(
-        prompt2blog_routes,
+        prompt2blog_runs,
         "read_status",
         lambda _run_id: {"run_id": _run_id, "feature": "prompt2blog"},
     )
     monkeypatch.setattr(
-        prompt2blog_routes,
+        prompt2blog_runs,
         "read_output",
         lambda _run_id: {
             "markdown": "# Prompt2Blog",
@@ -35,11 +37,11 @@ def test_prompt2blog_result_includes_langgraph_trace_metadata(monkeypatch):
         },
     )
     monkeypatch.setattr(
-        prompt2blog_routes,
+        prompt2blog_observability,
         "read_stage_result",
-        lambda _run_id, stage: {"data": trace_payload}
-        if stage == "langgraph_trace"
-        else None,
+        lambda _run_id, stage: (
+            {"data": trace_payload} if stage == "langgraph_trace" else None
+        ),
     )
 
     response = asyncio.run(prompt2blog_routes.get_result(run_id))
@@ -85,9 +87,9 @@ def test_url2blog_result_includes_langgraph_trace_metadata(monkeypatch):
     monkeypatch.setattr(
         url2blog_routes,
         "read_stage_result",
-        lambda _run_id, stage: {"data": trace_payload}
-        if stage == "langgraph_trace"
-        else None,
+        lambda _run_id, stage: (
+            {"data": trace_payload} if stage == "langgraph_trace" else None
+        ),
     )
 
     response = asyncio.run(url2blog_routes.get_result(run_id))
