+++ apps/ai-blog-writer/apps/backend/tests/test_prompt2blog_graph_routing.py
@@ -8,23 +8,23 @@ utils_stub.get_vertex_llm = lambda *args, **kwargs: None
 utils_stub.parse_json_response = lambda value: json.loads(value)
 sys.modules.setdefault("utils", utils_stub)
 
-import app.features.prompt2blog.routes as prompt2blog_routes
+import app.features.prompt2blog.orchestrator as orchestrator  # noqa: E402
+from app.features.prompt2blog.models import (  # noqa: E402
+    PipelineV2RuntimeRequest,
+    Prompt2BlogInputRequest,
+)
 
 
-def test_prompt2blog_pipeline_v2_routes_through_graph_runner(monkeypatch):
+def test_prompt2blog_pipeline_v2_routes_each_stage_through_graph(monkeypatch):
     captured: dict[str, object] = {}
 
-    def _fake_graph_runner(*, run_id, pipeline_runner):
-        captured["run_id"] = run_id
-        captured["pipeline_runner"] = pipeline_runner
+    def _fake_graph_runner(**kwargs):  # noqa: ANN003
+        captured.update(kwargs)
+        return kwargs["initial_state"]
 
-    monkeypatch.setattr(
-        prompt2blog_routes,
-        "run_prompt2blog_pipeline_v2_graph",
-        _fake_graph_runner,
-    )
+    monkeypatch.setattr(orchestrator, "run_prompt2blog_stage_graph", _fake_graph_runner)
 
-    request = prompt2blog_routes.PipelineV2RuntimeRequest(
+    request = PipelineV2RuntimeRequest(
         cleaned_data="Cleaned source",
         article_type_id=3,
         raw_sources=["source"],
@@ -33,28 +33,32 @@ def test_prompt2blog_pipeline_v2_routes_through_graph_runner(monkeypatch):
         include_debug=False,
         enable_editorial_augmentation=False,
     )
-
-    prompt2blog_routes._run_pipeline_v2("run-p2b-v2", request)
+    orchestrator.run_pipeline_v2("run-p2b-v2", request)
 
     assert captured["run_id"] == "run-p2b-v2"
-    assert callable(captured["pipeline_runner"])
+    assert [name for name, _node in captured["nodes"]] == [
+        "guideline",
+        "coverage",
+        "supplement",
+        "compose",
+        "quality_audit",
+        "repair",
+        "editorial_augmentation",
+        "title",
+        "finalize",
+    ]
 
 
-def test_prompt2blog_full_run_routes_through_graph_runner(monkeypatch):
+def test_prompt2blog_full_run_prepends_preparation_node(monkeypatch):
     captured: dict[str, object] = {}
 
-    def _fake_graph_runner(*, run_id, prepare_runner, pipeline_runner):
-        captured["run_id"] = run_id
-        captured["prepare_runner"] = prepare_runner
-        captured["pipeline_runner"] = pipeline_runner
+    def _fake_graph_runner(**kwargs):  # noqa: ANN003
+        captured.update(kwargs)
+        return kwargs["initial_state"]
 
-    monkeypatch.setattr(
-        prompt2blog_routes,
-        "run_prompt2blog_full_graph",
-        _fake_graph_runner,
-    )
+    monkeypatch.setattr(orchestrator, "run_prompt2blog_stage_graph", _fake_graph_runner)
 
-    request = prompt2blog_routes.Prompt2BlogInputRequest(
+    request = Prompt2BlogInputRequest(
         article_type_id=3,
         source_material=["raw source"],
         article_goal="Draft a practical article.",
@@ -65,9 +69,18 @@ def test_prompt2blog_full_run_routes_through_graph_runner(monkeypatch):
         include_debug=False,
         enable_editorial_augmentation=False,
     )
-
-    prompt2blog_routes._run_full_pipeline("run-p2b-full", request)
+    orchestrator.run_full_pipeline("run-p2b-full", request)
 
     assert captured["run_id"] == "run-p2b-full"
-    assert callable(captured["prepare_runner"])
-    assert callable(captured["pipeline_runner"])
+    assert [name for name, _node in captured["nodes"]] == [
+        "prepare",
+        "guideline",
+        "coverage",
+        "supplement",
+        "compose",
+        "quality_audit",
+        "repair",
+        "editorial_augmentation",
+        "title",
+        "finalize",
+    ]
